\name{head.DEresults}
\alias{head.DEresults}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Preview of Data in \code{DEresults} object
}
\description{
Modification of the \code{\link{head}} method for objects of type \code{DEresults}.
}
\usage{
\method{head}{DEresults}(x, \dots)
}
\arguments{
  \item{x}{
An item of type \code{DEresults}
}
  \item{\dots}{
Additional parameters to be passed to \code{\link{head}}
}
}

\author{
Jonas Mueller
}

\seealso{
\code{\link{head}}
}
