\name{head.plmDE}
\alias{head.plmDE}
\title{
Preview of \code{plmDE} object
}
\description{
Modification of \code{\link{head}} for objects of type \code{plmDE}.
}

\usage{
\method{head}{plmDE}(x, \dots)
}
\arguments{
  \item{x}{
The object to preview.
}
  \item{\dots}{
Additional parameters to be passed to \code{\link{head}}
}
}

\author{
Jonas Mueller
}

\seealso{
\code{\link{head}}
}
