\name{info}
\alias{info}
\title{
Information about an Object
}
\description{
Prints out a short summary of the data contained with an Object.
}
\usage{
info(Object, ...)
}
\arguments{
 \item{Object}{
print information about \code{Object}
}
 \item{\dots}{
additional instructions}
}

\author{
Jonas Mueller}

\seealso{
\code{\link{info.plmDE}}
}

\keyword{ ~print }