\name{plmDEmodel}
\alias{plmDEmodel}
\title{
Constructor for objects of type \code{plmDE}
}
\description{
Creates an object which can hold expression data and other measurements of interest.
}
\usage{
plmDEmodel(genes, \dots)
}

\arguments{
  \item{genes}{
Object containing the genes of interest from which an object of type \code{plmDE} should be made. 
}
 \item{\dots}{
additional parameters}
}

\author{
Jonas Mueller
}

\seealso{
\code{\link{plmDEmodel.default}}
}
